feat: move test files and fix them Refs: #49
